testing/web-platform/tests/workers/WorkerGlobalScope_XMLHttpRequest.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerGlobalScope XMLHttpRequest() support </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var FileName = './support/XMLHttpRequest.js';
    var RequstUrl = 'XMLHttpRequest.txt';
    var ExpectedResult = "Request Succeeded!";

    var t = async_test("Test Description: WorkerGlobalScope XMLHttpRequest() support.");
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_equals(evt.data, ExpectedResult);
        t.done();
    });

    worker.postMessage(RequstUrl);
</script>
</body>
</html>
